requirements.txt
streamlit>=1.38
anthropic>=0.45
pymupdf>=1.24
python-dotenv>=1.0
